BlockChains/requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.18.0
networkx>=3.1
python-dotenv>=0.19.0

# Blockchain dependencies
flask>=2.0.1
requests>=2.26.0
gunicorn>=20.1.0

# Data visualization
matplotlib>=3.7.0
seaborn>=0.12.0 